// src/generated/api/models/capabilitiesResponseSearch.ts
export interface CapabilitiesResponseSearch {
    /**
     * 
     * @type {number}
     * @memberof CapabilitiesResponseSearch
     */
    defaultRadiusM: number;
    /**
     * 
     * @type {number}
     * @memberof CapabilitiesResponseSearch
     */
    minRadiusM: number;
    /**
     * 
     * @type {number}
     * @memberof CapabilitiesResponseSearch
     */
    maxRadiusM: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof CapabilitiesResponseSearch
     */
    radiusPresetsM: Array<number>;
    /**
     * 
     * @type {number}
     * @memberof CapabilitiesResponseSearch
     */
    defaultPageSize: number;
    /**
     * 
     * @type {number}
     * @memberof CapabilitiesResponseSearch
     */
    maxPageSize: number;
    /**
     * 
     * @type {PaginationMode}
     * @memberof CapabilitiesResponseSearch
     */
    paginationMode: PaginationMode;
    /**
     * 
     * @type {Array<ProviderMode>}
     * @memberof CapabilitiesResponseSearch
     */
    providerModes: Array<ProviderMode>;
    /**
     * 
     * @type {Array<RestaurantSortBy>}
     * @memberof CapabilitiesResponseSearch
     */
    sortBy: Array<RestaurantSortBy>;
    /**
     * 
     * @type {Array<SortDirection>}
     * @memberof CapabilitiesResponseSearch
     */
    sortDirections: Array<SortDirection>;
    /**
     * 
     * @type {Array<RestaurantFeature>}
     * @memberof CapabilitiesResponseSearch
     */
    requiredFeatures: Array<RestaurantFeature>;
}